+++ index.php
@@ -11,6 +11,11 @@ if(isset($_GET['nux-mac']) && !empty($_GET['nux-mac'])){
 if(isset($_GET['nux-ip']) && !empty($_GET['nux-ip'])){
     $_SESSION['nux-ip'] = $_GET['nux-ip'];
 }
+
+if(isset($_GET['nux-router']) && !empty($_GET['nux-router'])){
+    $_SESSION['nux-router'] = $_GET['nux-router'];
+}
+
 require_once 'system/vendor/autoload.php';
 require_once 'system/boot.php';
 App::_run();