+++ CHANGELOG.md
@@ -2,12 +2,20 @@
 
 # CHANGELOG
 
-## 2023.9.15
+## 2023.9.21
+
+- Customer can extend Plan
+- Customer can Deactivate active plan
+- add variable nux-router to select  only plan from that router
+- Show user expired until 30 items
+
+## 2023.9.20
 
 - Fix Customer balance header
 - Deactivate Customer active plan
 - Sync Customer Plan to Mikrotik
 - Recharge Customer from Customer Details
+- Add Privacy Policy and Terms and Conditions Pages
 
 ## 2023.9.13